+++ app/test_helpers.go
@@ -12,7 +12,6 @@ import (
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 	tmtypes "github.com/tendermint/tendermint/types"
 	dbm "github.com/tendermint/tm-db"
-	feemarkettypes "github.com/tharsis/ethermint/x/feemarket/types"
 )
 
 // DefaultConsensusParams defines the default Tendermint consensus params used in
@@ -35,20 +34,14 @@ var DefaultConsensusParams = &abci.ConsensusParams{
 }
 
 // Setup initializes a new EthermintApp. A Nop logger is set in EthermintApp.
-func Setup(isCheckTx bool, feemarketGenesis *feemarkettypes.GenesisState) *EthermintApp {
+func Setup(isCheckTx bool, patchGenesis func(*EthermintApp, simapp.GenesisState) simapp.GenesisState) *EthermintApp {
 	db := dbm.NewMemDB()
 	app := NewEthermintApp(log.NewNopLogger(), db, nil, true, map[int64]bool{}, DefaultNodeHome, 5, encoding.MakeConfig(ModuleBasics), simapp.EmptyAppOptions{})
 	if !isCheckTx {
 		// init chain must be called to stop deliverState from being nil
 		genesisState := NewDefaultGenesisState()
-
-		// Verify feeMarket genesis
-		if feemarketGenesis != nil {
-			if err := feemarketGenesis.Validate(); err != nil {
-				panic(err)
-			}
-
-			genesisState[feemarkettypes.ModuleName] = app.AppCodec().MustMarshalJSON(feemarketGenesis)
+		if patchGenesis != nil {
+			genesisState = patchGenesis(app, genesisState)
 		}
 
 		stateBytes, err := json.MarshalIndent(genesisState, "", " ")